src/phase_3_2_retrieval/requirements.txt
langchain
langchain-community
chromadb
sentence-transformers
langchain-groq
python-dotenv
